+++ src/ShmsBackend.Api/Services/Reward/IRewardService.cs
@@ -12,5 +12,5 @@ public class RedeemPointsResult
 public interface IRewardService
 {
     Task EarnPointsAsync(Guid tenantId, Guid houseId, decimal amountReceived, bool isInitialPayment, Guid relatedPaymentId);
-    Task<RedeemPointsResult> RedeemPointsAsync(Guid tenantId, int pointsToRedeem);
+    Task<RedeemPointsResult> RedeemPointsAsync(Guid tenantId, decimal pointsToRedeem);
 }
